requirements.txt
django==4.2.5
djangorestframework==3.14.0
cryptography==41.0.4